v0 of drawing support for CircularString geometries
