+++ src/ol/geom/CircularString.js
@@ -1,18 +1,17 @@
 /**
  * @module ol/geom/CircularString
  */
-import GeometryLayout from './GeometryLayout.js';
 import GeometryType from './GeometryType.js';
 import SimpleGeometry from './SimpleGeometry.js';
 import {assignClosestPoint, maxSquaredDelta} from './flat/closest.js';
-import {closestSquaredDistanceXY} from '../extent.js';
+import {
+  closestSquaredDistanceXY,
+  createEmpty,
+  extendFlatCoordinates,
+  intersects,
+} from '../extent.js';
 import {deflateCoordinates} from './flat/deflate.js';
-import {extend} from '../array.js';
-import {forEach as forEachSegment} from './flat/segments.js';
 import {inflateCoordinates} from './flat/inflate.js';
-import {interpolatePoint, lineStringCoordinateAtM} from './flat/interpolate.js';
-import {intersectsLineString} from './flat/intersectsextent.js';
-import {lineStringLength} from './flat/length.js';
 
 /**
  * @classdesc
@@ -28,18 +27,6 @@ class CircularString extends SimpleGeometry {
    */
   constructor(coordinates, opt_layout) {
     super();
-
-    /**
-     * @private
-     * @type {import("../coordinate.js").Coordinate}
-     */
-    this.flatMidpoint_ = null;
-
-    /**
-     * @private
-     * @type {number}
-     */
-    this.flatMidpointRevision_ = -1;
 
     /**
      * @private
@@ -69,20 +56,6 @@ class CircularString extends SimpleGeometry {
   }
 
   /**
-   * Append the passed coordinate to the coordinates of the linestring.
-   * @param {import("../coordinate.js").Coordinate} coordinate Coordinate.
-   * @api
-   */
-  appendCoordinate(coordinate) {
-    if (!this.flatCoordinates) {
-      this.flatCoordinates = coordinate.slice();
-    } else {
-      extend(this.flatCoordinates, coordinate);
-    }
-    this.changed();
-  }
-
-  /**
    * Make a complete copy of the geometry.
    * @return {!CircularString} Clone.
    * @api
@@ -107,7 +80,7 @@ class CircularString extends SimpleGeometry {
     if (minSquaredDistance < closestSquaredDistanceXY(this.getExtent(), x, y)) {
       return minSquaredDistance;
     }
-    if (this.maxDeltaRevision_ != this.getRevision()) {
+    if (this.maxDeltaRevision_ !== this.getRevision()) {
       this.maxDelta_ = Math.sqrt(
         maxSquaredDelta(
           this.flatCoordinates,
@@ -134,60 +107,7 @@ class CircularString extends SimpleGeometry {
   }
 
   /**
-   * Iterate over each segment, calling the provided callback.
-   * If the callback returns a truthy value the function returns that
-   * value immediately. Otherwise the function returns `false`.
-   *
-   * @param {function(this: S, import("../coordinate.js").Coordinate, import("../coordinate.js").Coordinate): T} callback Function
-   *     called for each segment. The function will receive two arguments, the start and end coordinates of the segment.
-   * @return {T|boolean} Value.
-   * @template T,S
-   * @api
-   */
-  forEachSegment(callback) {
-    return forEachSegment(
-      this.flatCoordinates,
-      0,
-      this.flatCoordinates.length,
-      this.stride,
-      callback
-    );
-  }
-
-  /**
-   * Returns the coordinate at `m` using linear interpolation, or `null` if no
-   * such coordinate exists.
-   *
-   * `opt_extrapolate` controls extrapolation beyond the range of Ms in the
-   * MultiLineString. If `opt_extrapolate` is `true` then Ms less than the first
-   * M will return the first coordinate and Ms greater than the last M will
-   * return the last coordinate.
-   *
-   * @param {number} m M.
-   * @param {boolean} [opt_extrapolate] Extrapolate. Default is `false`.
-   * @return {import("../coordinate.js").Coordinate} Coordinate.
-   * @api
-   */
-  getCoordinateAtM(m, opt_extrapolate) {
-    if (
-      this.layout != GeometryLayout.XYM &&
-      this.layout != GeometryLayout.XYZM
-    ) {
-      return null;
-    }
-    const extrapolate = opt_extrapolate !== undefined ? opt_extrapolate : false;
-    return lineStringCoordinateAtM(
-      this.flatCoordinates,
-      0,
-      this.flatCoordinates.length,
-      this.stride,
-      m,
-      extrapolate
-    );
-  }
-
-  /**
-   * Return the coordinates of the linestring.
+   * Return the coordinates of the circular string.
    * @return {Array<import("../coordinate.js").Coordinate>} Coordinates.
    * @api
    */
@@ -198,53 +118,6 @@ class CircularString extends SimpleGeometry {
       this.flatCoordinates.length,
       this.stride
     );
-  }
-
-  /**
-   * Return the coordinate at the provided fraction along the linestring.
-   * The `fraction` is a number between 0 and 1, where 0 is the start of the
-   * linestring and 1 is the end.
-   * @param {number} fraction Fraction.
-   * @param {import("../coordinate.js").Coordinate} [opt_dest] Optional coordinate whose values will
-   *     be modified. If not provided, a new coordinate will be returned.
-   * @return {import("../coordinate.js").Coordinate} Coordinate of the interpolated point.
-   * @api
-   */
-  getCoordinateAt(fraction, opt_dest) {
-    return interpolatePoint(
-      this.flatCoordinates,
-      0,
-      this.flatCoordinates.length,
-      this.stride,
-      fraction,
-      opt_dest,
-      this.stride
-    );
-  }
-
-  /**
-   * Return the length of the linestring on projected plane.
-   * @return {number} Length (on projected plane).
-   * @api
-   */
-  getLength() {
-    return lineStringLength(
-      this.flatCoordinates,
-      0,
-      this.flatCoordinates.length,
-      this.stride
-    );
-  }
-
-  /**
-   * @return {Array<number>} Flat midpoint.
-   */
-  getFlatMidpoint() {
-    if (this.flatMidpointRevision_ != this.getRevision()) {
-      this.flatMidpoint_ = this.getCoordinateAt(0.5, this.flatMidpoint_);
-      this.flatMidpointRevision_ = this.getRevision();
-    }
-    return this.flatMidpoint_;
   }
 
   /**
@@ -263,17 +136,25 @@ class CircularString extends SimpleGeometry {
    * @api
    */
   intersectsExtent(extent) {
-    return intersectsLineString(
-      this.flatCoordinates,
-      0,
-      this.flatCoordinates.length,
-      this.stride,
-      extent
-    );
+    // test per arc extents
+    for (let i = 0; i < this.flatCoordinates.length - 2; i += 4) {
+      const arcExtents = extendFlatCoordinates(
+        createEmpty(),
+        this.flatCoordinates,
+        i,
+        i + this.stride * 3,
+        this.stride
+      );
+      if (intersects(extent, arcExtents)) {
+        return true;
+      }
+    }
+
+    return false;
   }
 
   /**
-   * Set the coordinates of the linestring.
+   * Set the coordinates of the circular string.
    * @param {!Array<import("../coordinate.js").Coordinate>} coordinates Coordinates.
    * @param {import("./GeometryLayout.js").default} [opt_layout] Layout.
    * @api
