+++ src/ol/render/canvas/CircularStringBuilder.js
@@ -6,7 +6,6 @@ import CanvasInstruction, {
   beginPathInstruction,
   strokeInstruction,
 } from './Instruction.js';
-import {defaultLineDash, defaultLineDashOffset} from '../canvas.js';
 
 class CanvasCircularStringBuilder extends CanvasBuilder {
   /**
@@ -21,32 +20,22 @@ class CanvasCircularStringBuilder extends CanvasBuilder {
 
   /**
    * @param {Array<number>} flatCoordinates Flat coordinates.
-   * @param {number} offset Offset.
-   * @param {number} end End.
-   * @param {number} stride Stride.
-   * @private
-   * @return {number} end.
+   * @param {number} stride The stride.
    */
-  drawFlatCoordinates_(flatCoordinates, offset, end, stride) {
-    const myBegin = this.coordinates.length;
-    const myEnd = this.appendFlatLineCoordinates(
-      flatCoordinates,
-      offset,
-      end,
-      stride,
-      false,
-      false
-    );
-    const moveToLineToInstruction = [
-      CanvasInstruction.MOVE_TO_LINE_TO,
-      myBegin,
-      myEnd,
-    ];
-    this.instructions.push(moveToLineToInstruction);
-    this.hitDetectionInstructions.push(moveToLineToInstruction);
-    return end;
+  appendFlatCircularStringCoordinates(flatCoordinates, stride) {
+    const coordinates = this.coordinates;
+    const length = flatCoordinates.length;
+    let myEnd = this.coordinates.length;
+    for (let i = 0; i < length; i += stride) {
+      coordinates[myEnd++] = flatCoordinates[i];
+      coordinates[myEnd++] = flatCoordinates[i + 1];
+    }
   }
 
+  /**
+   * @param {import("../../geom/CircularString.js").default|import("../Feature.js").default} circularStringGeometry Circular string geometry.
+   * @param {import("../../Feature.js").FeatureLike} feature Feature.
+   */
   drawCircularString(circularStringGeometry, feature) {
     const state = this.state;
     const strokeStyle = state.strokeStyle;
@@ -55,104 +44,16 @@ class CanvasCircularStringBuilder extends CanvasBuilder {
       return;
     }
     this.updateStrokeStyle(state, this.applyStroke);
-
-    //console.log('Draw "CircularString"');
-
     this.beginGeometry(circularStringGeometry, feature);
     const startIndex = this.coordinates.length;
     const coordinates = circularStringGeometry.getFlatCoordinates();
     const stride = circularStringGeometry.getStride();
-    this.appendFlatPointCoordinates(coordinates, stride);
-
+    this.appendFlatCircularStringCoordinates(coordinates, stride);
     this.instructions.push([
       CanvasInstruction.CIRCULAR_ARC,
       startIndex,
       startIndex + coordinates.length,
     ]);
-
-    this.endGeometry(feature);
-
-    // console.log('Draw "LineString"');
-    // this.drawLineString(circularStringGeometry, feature);
-  }
-
-  /**
-   * @param {import("../../geom/LineString.js").default|import("../Feature.js").default} lineStringGeometry Line string geometry.
-   * @param {import("../../Feature.js").FeatureLike} feature Feature.
-   */
-  drawLineString(lineStringGeometry, feature) {
-    const state = this.state;
-    const strokeStyle = state.strokeStyle;
-    const lineWidth = state.lineWidth;
-    if (strokeStyle === undefined || lineWidth === undefined) {
-      return;
-    }
-    this.updateStrokeStyle(state, this.applyStroke);
-    this.beginGeometry(lineStringGeometry, feature);
-    this.hitDetectionInstructions.push(
-      [
-        CanvasInstruction.SET_STROKE_STYLE,
-        state.strokeStyle,
-        state.lineWidth,
-        state.lineCap,
-        state.lineJoin,
-        state.miterLimit,
-        defaultLineDash,
-        defaultLineDashOffset,
-      ],
-      beginPathInstruction
-    );
-    const flatCoordinates = lineStringGeometry.getFlatCoordinates();
-    const stride = lineStringGeometry.getStride();
-    this.drawFlatCoordinates_(
-      flatCoordinates,
-      0,
-      flatCoordinates.length,
-      stride
-    );
-    this.hitDetectionInstructions.push(strokeInstruction);
-    this.endGeometry(feature);
-  }
-
-  /**
-   * @param {import("../../geom/MultiLineString.js").default|import("../Feature.js").default} multiLineStringGeometry MultiLineString geometry.
-   * @param {import("../../Feature.js").FeatureLike} feature Feature.
-   */
-  drawMultiLineString(multiLineStringGeometry, feature) {
-    const state = this.state;
-    const strokeStyle = state.strokeStyle;
-    const lineWidth = state.lineWidth;
-    if (strokeStyle === undefined || lineWidth === undefined) {
-      return;
-    }
-    this.updateStrokeStyle(state, this.applyStroke);
-    this.beginGeometry(multiLineStringGeometry, feature);
-    this.hitDetectionInstructions.push(
-      [
-        CanvasInstruction.SET_STROKE_STYLE,
-        state.strokeStyle,
-        state.lineWidth,
-        state.lineCap,
-        state.lineJoin,
-        state.miterLimit,
-        state.lineDash,
-        state.lineDashOffset,
-      ],
-      beginPathInstruction
-    );
-    const ends = multiLineStringGeometry.getEnds();
-    const flatCoordinates = multiLineStringGeometry.getFlatCoordinates();
-    const stride = multiLineStringGeometry.getStride();
-    let offset = 0;
-    for (let i = 0, ii = ends.length; i < ii; ++i) {
-      offset = this.drawFlatCoordinates_(
-        flatCoordinates,
-        offset,
-        /** @type {number} */ (ends[i]),
-        stride
-      );
-    }
-    this.hitDetectionInstructions.push(strokeInstruction);
     this.endGeometry(feature);
   }
